move X25519 tests to their own file
